Topic refactored. Age category. StoryDetails cleanup.

+++ XooCreator.BA/Data/Entities/CreditWallet.cs
@@ -5,8 +5,8 @@ public class CreditWallet
     public Guid UserId { get; set; }
     public AlchimaliaUser User { get; set; } = null!;
     // Generative credits (legacy Balance kept for backwards compatibility)
-    public int Balance { get; set; }
+    public double Balance { get; set; }
     // New: Discovery credits balance
-    public int DiscoveryBalance { get; set; }
+    public double DiscoveryBalance { get; set; }
     public DateTime UpdatedAt { get; set; } = DateTime.UtcNow;
 }
